Implementation of view model
