+++ app/src/main/java/com/angelemv/android/pruebatecnicasps/views/MainScreen.kt
@@ -13,27 +13,24 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.lifecycle.viewmodel.compose.viewModel
 import coil.compose.AsyncImage
 import com.angelemv.android.pruebatecnicasps.model.data.User
 import com.angelemv.android.pruebatecnicasps.model.interfaces.RetrofitInstance.apiService
 import com.angelemv.android.pruebatecnicasps.ui.theme.PruebaTecnicaSPSTheme
+import com.angelemv.android.pruebatecnicasps.viewmodel.MainViewModel
 
 @Composable
-fun MainScreen() {
-    var users by remember { mutableStateOf<List<User>>(emptyList()) }
-    LaunchedEffect(Unit) {
-        users = api()
-    }
+fun MainScreen(mainViewModel: MainViewModel = viewModel()) {
+    val users by mainViewModel.users.collectAsState()
+
     if (users.isEmpty()) {
         Column(
             modifier = Modifier.fillMaxSize(),
@@ -43,7 +40,8 @@ fun MainScreen() {
             Text("Cargando usuarios...")
         }
     } else {
-        LazyColumn(modifier = Modifier.fillMaxSize()
+        LazyColumn(modifier = Modifier
+            .fillMaxSize()
             .padding(top = 32.dp)
         ) {
             items(users) { user ->
@@ -74,16 +72,6 @@ fun UserItem(user: User) {
             Text(text = "Apellido: ${user.last_name}")
             Text(text = "Email: ${user.email}")
         }
-    }
-}
-
-suspend fun api(): List<User> {
-    return try {
-        val response = apiService.getUsers()
-        response.data
-    } catch (e: Exception) {
-        e.printStackTrace()
-        emptyList()
     }
 }
